+++ src/models/settings/CellSettings.java
@@ -3,25 +3,25 @@ package models.settings;
 import java.util.HashMap;
 
 public class CellSettings {
-	private HashMap<Integer, HashMap<String, Integer>> fProperties;
+	private HashMap<Integer, HashMap<String, Double>> fProperties;
 	private HashMap<Integer, String> fRuleName;
 	private HashMap<Integer, String> fNeighborhoodName;
 	
 	public CellSettings()
 	{
-		fProperties = new HashMap<Integer, HashMap<String, Integer>>();
+		fProperties = new HashMap<Integer, HashMap<String, Double>>();
 		fRuleName = new HashMap<Integer, String>();
 		fNeighborhoodName = new HashMap<Integer, String>();
 	}
 	
-	public void addProperty(int aStateIndex, String aKey, int aValue)
+	public void addProperty(int aStateIndex, String aKey, double aValue)
 	{
 		if (!fProperties.containsKey(aStateIndex)) {
-			fProperties.put(aStateIndex, new HashMap<String,Integer>());
+			fProperties.put(aStateIndex, new HashMap<String,Double>());
 		}
 		fProperties.get(aStateIndex).put(aKey, aValue);
 	}
-	public HashMap<String, Integer> getProperties(int aStateIndex)
+	public HashMap<String, Double> getProperties(int aStateIndex)
 	{
 		return fProperties.get(aStateIndex);
 	}
